+++ R/get_geo_peru.R
@@ -4,9 +4,10 @@
 #' (EPSG:4326).
 #'
 #' @param geography A character or a vector with the name of geographical region.
-#' An exception is "all" to request all Peru data.
-#' @param level A character: "all" for national level data, "dep" for department
-#' level data, and "prov" for provincial level data.
+#' Use "all" to request all geographic units at the selected level.
+#' @param level A character: "all" for national data, "dep" for departments,
+#' and "prov" for provinces. Together with `geography = "all"`, the options
+#' return one national dataset, 25 departments, or 196 provinces, respectively.
 #' @param simplified A logical TRUE or FALSE, to select data with all districts or
 #' a polygon simplified to a higher level.
 #' @param showProgress Logical TRUE or FALSE to display a progress bar during download.